requirements.txt
# ISEE Tutor - Complete Requirements
# For NVIDIA Jetson Orin Nano with Ubuntu

# Core Web Framework
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
websockets==12.0
aiofiles==23.2.1

# API and Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1
alembic==1.13.1

# AI/ML Core
langchain==0.1.6
langchain-community==0.0.19
transformers==4.37.2
sentence-transformers==2.3.1
# Note: For Jetson, install PyTorch separately with:
# pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu

# LLM Support
llama-cpp-python==0.2.38
huggingface-hub==0.20.3

# Vector Database
chromadb==0.4.22
tiktoken==0.5.2

# Speech Recognition & TTS
openai-whisper==20231117
SpeechRecognition==3.10.1
pyaudio==0.2.14
pyttsx3==2.90
TTS==0.22.0
webrtcvad==2.0.10
sounddevice==0.4.6
soundfile==0.12.1

# Audio Processing
librosa==0.10.1
pydub==0.25.1
scipy==1.12.0
numpy==1.24.3

# Document Processing
pypdf2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
python-docx==1.1.0
beautifulsoup4==4.12.3
easyocr==1.7.0

# Hardware Control (Jetson specific)
# Jetson.GPIO is pre-installed on Jetson
pyserial==3.5
rpi-ws281x==5.0.0  # For LED control

# Environment and Config
python-dotenv==1.0.1
pyyaml==6.0.1
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
click==8.1.7
requests==2.31.0
httpx==0.26.0
tqdm==4.66.1
colorama==0.4.6
python-dateutil==2.8.2

# Development and Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
ipython==8.20.0
jupyter==1.0.0

# Monitoring and Logging
prometheus-client==0.19.0
structlog==24.1.0

# Image Processing (for touchscreen UI)
pillow==10.2.0
opencv-python==4.9.0.80

# Data Science
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.1

# Background Tasks
celery==5.3.4
flower==2.0.1

# Object Storage
minio==7.2.3

# Additional ML Tools
scikit-learn==1.4.0
xgboost==2.0.3

# Caching
diskcache==5.6.3

# Schema Validation
jsonschema==4.21.1
marshmallow==3.20.2